Remove image operation commands
