+++ bot/utilities.go
@@ -16,6 +16,8 @@ type ImageOperationArgs interface {
 	GetImageURL() string
 }
 
+type ImageOperation[K ImageOperationArgs] func([]byte, io.Writer, K) error
+
 // TypingIndicator invokes a typing indicator in the channel of a message
 func TypingIndicator(message *discordgo.MessageCreate) func() {
 	stopTyping := Schedule(
@@ -32,6 +34,24 @@ func TypingIndicator(message *discordgo.MessageCreate) func() {
 	return func() {
 		stopTyping <- true
 	}
+}
+
+// Schedule some func to be run in a cancelable goroutine on an interval
+func Schedule(what func(), delay time.Duration) chan bool {
+	stop := make(chan bool)
+
+	go func() {
+		for {
+			what()
+			select {
+			case <-time.After(delay):
+			case <-stop:
+				return
+			}
+		}
+	}()
+
+	return stop
 }
 
 // ImageURLFromMessage attempts to retrieve an image URL from a given message.
@@ -71,7 +91,7 @@ func FindImageURL(m *discordgo.MessageCreate) (string, error) {
 	return "", errors.New("unable to locate an image")
 }
 
-// DownloadImage downloads an image from a given URL, returing the resulting bytes.
+// DownloadImage downloads an image from a given URL, returning the resulting bytes.
 func DownloadImage(url string) ([]byte, error) {
 	log.Debug().Str("url", url).Msg("Downloading image")
 	resp, err := http.Get(url)
@@ -90,26 +110,15 @@ func DownloadImage(url string) ([]byte, error) {
 	return buffer.Bytes(), nil
 }
 
-// Schedule some func to be ran in a cancelable goroutine on an interval
-func Schedule(what func(), delay time.Duration) chan bool {
-	stop := make(chan bool)
-
-	go func() {
-		for {
-			what()
-			select {
-			case <-time.After(delay):
-			case <-stop:
-				return
-			}
-		}
-	}()
-
-	return stop
+// MakeImageOpCommand automatically creates a Parsley command handler for a given ImageOperation
+func MakeImageOpCommand[K ImageOperationArgs](operation ImageOperation[K]) func(*discordgo.MessageCreate, K) {
+	return func(message *discordgo.MessageCreate, args K) {
+		PrepareAndInvokeOperation(message, args, operation)
+	}
 }
 
-// PrepareAndInvokeOperation downloads the image pulled from the message, invokes the given operation with said image, and posts the image in the channel of the message that invoked it
-func PrepareAndInvokeOperation[K ImageOperationArgs](message *discordgo.MessageCreate, args K, operation func([]byte, io.Writer, K) error) {
+// PrepareAndInvokeOperation automatically handles invoking a given ImageOperation for a Discord message and returning the finished results.
+func PrepareAndInvokeOperation[K ImageOperationArgs](message *discordgo.MessageCreate, args K, operation ImageOperation[K]) {
 	defer TypingIndicator(message)()
 
 	imageUrl := args.GetImageURL()
